requirements.txt
pytrends
requests
python-telegram-bot==13.15